Standardizes deployment code and cleans it up making it a little easier to add AWS when the time comes
