+++ .circleci/config.yml
@@ -38,29 +38,18 @@ jobs:
   build_push:
     machine:
       docker_layer_caching: true
-    working_directory: ~/repo
-
     steps:
       - checkout
       - run:
-          name: Setup Environment Variables
-          command: |
-            echo 'export IMAGE_REGISTRY="${PROJECT_REGISTRY_URL:-gcr.io/${PROJECT_ID}/}"' >> $BASH_ENV
-            if [[ $CIRCLE_BRANCH == "master" ]]; then
-              echo 'export IMAGE_TAG="latest"' >> $BASH_ENV
-            else
-              echo 'export IMAGE_TAG="dev"' >> $BASH_ENV
-            fi
-      - run:
           name: Build Container
           command: |
-            docker login -u _json_key -p "${GCLOUD_SERVICE_KEY}" https://gcr.io
-            docker compose build --push deploy-gce
+            cd /sv/
+            bash scripts/run_ci.sh
       - notify-channel:
           only-on-fail: true
 
   complete:
     docker:
-      - image: cimg/node:lts
+      - image: cimg/node:20.19.4
     steps:
       - notify-channel
