--- deploy/gce/lib/exec.js
@@ -1,7 +1,0 @@
-const child_process = require("child_process");
-
-function exec(command) {
-	return child_process.execSync(command, { stdio : "inherit" });
-}
-
-module.exports = exec;
