--- deploy/gce/scripts/build_external
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-const buildExternal = require("../lib/buildExternal");
-
-buildExternal();
