--- deploy/gce/scripts/setup
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-const setup = require("../lib/setup");
-
-setup();
