--- deploy/gce/scripts/test
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-const test = require("../lib/test");
-
-test();
