--- deploy/gce/scripts/undeploy
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-const undeploy = require("../lib/undeploy");
-
-undeploy();
