+++ docker-compose.yml
@@ -15,23 +15,21 @@ services:
       - "./internal/gcloud:/root/.config/gcloud"
       - "//var/run/docker.sock:/var/run/docker.sock"
 
-  deploy-gce:
-    image: ${IMAGE_REGISTRY:-}sv-deploy-gce:${IMAGE_TAG:-local}
+  deploy:
+    image: sv-deploy:local
     build:
       context: .
-      dockerfile: ./deploy/gce/Dockerfile
-      cache_from:
-        - ${IMAGE_REGISTRY:-}sv-deploy-gce:${IMAGE_TAG:-local}
+      dockerfile: ./Dockerfile.deploy
     init: true
     environment:
-      PROJECT_ID: sv-shared-231700
+      PROJECT_ID: sv-devops-231700
       REPO_NAME: sv-kubernetes-example
       BRANCH_NAME: develop
       ZONE: us-east1-b
-      GCLOUD_SERVICE_KEY: ${SERVICE_KEY:-}
     volumes:
-      - "./:/sv"
-      - "~/.kube:/root/.kube"
-      - "~/.ssh:/root/.ssh"
-      - "./internal/gcloud:/root/.config/gcloud"
+      - "./sv:/sv/sv"
+      - "./scripts:/sv/scripts"
+      - "./internal:/sv/internal"
+      - "./scripts/deploy:/app/scripts"
+      - "./applications/sv-kubernetes-example:/repo"
       - "//var/run/docker.sock:/var/run/docker.sock"
