+++ scripts/deploy/undeploy
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+const undeploy = require("/sv/sv/deploy/undeploy");
+
+undeploy();
